requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==3.1.1
attrs==23.2.0
certifi==2023.11.17
execnet==2.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
PySocks==1.7.1
pytest==7.4.4
pytest-xdist==3.5.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
